tietotyyppi muutettu suhteen laskemiseksi

+++ student/04/container/main.cpp
@@ -69,7 +69,7 @@ bool is_ordered_non_strict_ascending(const std::vector <int>& ints){
 bool is_geometric_series(const std::vector <int>& ints){
     int c = 0;
     int prev = 0;
-    double rel = ints.at(0) / ints.at(1);
+    double rel = static_cast <double> (ints.at(0)) / static_cast <double> (ints.at(1));
     for (int item : ints){
         if (item == 0){
             return false;
